fix: ignore deserialization error
