+++ src/pdm/models/session.py
@@ -8,12 +8,10 @@ from typing import TYPE_CHECKING, Any, cast
 
 import hishel
 import httpx
-import msgpack
-from hishel._serializers import Metadata
-from httpcore import Request, Response
 from unearth.fetchers import PyPIClient
 
 from pdm.__version__ import __version__
+from pdm.models.serializers import MsgPackSerializer
 from pdm.termui import logger
 
 if TYPE_CHECKING:
@@ -55,82 +53,6 @@ def _get_transport(
     proxy: httpx.Proxy | None = None,
 ) -> httpx.BaseTransport:
     return httpx.HTTPTransport(verify=verify, cert=cert, trust_env=True, proxy=proxy, retries=MAX_RETRIES)
-
-
-class MsgPackSerializer(hishel.BaseSerializer):
-    KNOWN_REQUEST_EXTENSIONS = ("timeout", "sni_hostname")
-    KNOWN_RESPONSE_EXTENSIONS = ("http_version", "reason_phrase")
-    DATETIME_FORMAT = "%Y-%m-%dT%H:%M:%S.%fZ"
-
-    def dumps(self, response: Response, request: Request, metadata: Metadata) -> bytes:
-        from hishel._utils import normalized_url
-
-        response_dict = {
-            "status": response.status,
-            "headers": response.headers,
-            "content": response.content,
-            "extensions": {
-                key: value for key, value in response.extensions.items() if key in self.KNOWN_RESPONSE_EXTENSIONS
-            },
-        }
-
-        request_dict = {
-            "method": request.method.decode("ascii"),
-            "url": normalized_url(request.url),
-            "headers": request.headers,
-            "extensions": {
-                key: value for key, value in request.extensions.items() if key in self.KNOWN_REQUEST_EXTENSIONS
-            },
-        }
-
-        metadata_dict = {
-            "cache_key": metadata["cache_key"],
-            "number_of_uses": metadata["number_of_uses"],
-            "created_at": metadata["created_at"].strftime(self.DATETIME_FORMAT),
-        }
-
-        full_dict = {
-            "response": response_dict,
-            "request": request_dict,
-            "metadata": metadata_dict,
-        }
-        return cast(bytes, msgpack.packb(full_dict, use_bin_type=True))
-
-    def loads(self, data: bytes) -> tuple[Response, Request, Metadata]:
-        from datetime import datetime
-
-        full_dict = cast("dict[str, Any]", msgpack.loads(data, raw=False))
-
-        response_dict = full_dict["response"]
-        request_dict = full_dict["request"]
-        metadata_dict = full_dict["metadata"]
-        metadata_dict["created_at"] = datetime.strptime(metadata_dict["created_at"], self.DATETIME_FORMAT)
-
-        response = Response(
-            status=response_dict["status"],
-            headers=response_dict["headers"],
-            content=response_dict["content"],
-            extensions=response_dict["extensions"],
-        )
-
-        request = Request(
-            method=request_dict["method"],
-            url=request_dict["url"],
-            headers=request_dict["headers"],
-            extensions=request_dict["extensions"],
-        )
-
-        metadata = Metadata(
-            cache_key=metadata_dict["cache_key"],
-            created_at=metadata_dict["created_at"],
-            number_of_uses=metadata_dict["number_of_uses"],
-        )
-
-        return response, request, metadata
-
-    @property
-    def is_binary(self) -> bool:
-        return True
 
 
 class PDMPyPIClient(PyPIClient):
